refactor(cli): one canonical formatBytes and compareVersions; cut 3 lib->commands imports
Structural scan of apps/cli/src found the same small utilities reimplemented
across the tree, and three lib/ modules importing upward into the command layer.

- formatBytes: five copies rendering the same number three different ways.
  agents prune showed 1.50 GB where every other surface showed 1.5 GB, and
  session artifact sizes capped at MB (1.4 GiB rendered as 1433.6 MB). One
  definition in lib/format.ts; its test moves there from inspect.test.ts.

- compareVersions: capabilities.ts and commands.ts each carried a private copy
  that dropped OpenClaw's -N rebuild-suffix tiebreak, so 2026.2.19-2 compared
  equal to 2026.2.19. Both now import the canonical comparator from
  agent-spec/primitives.ts, a leaf module -- the cycle the old comment cited as
  the reason for the copy never applied to it. No capability declares a -N floor
  today, so this fixes a latent gate, not a live one.

- isPromptCancelled / isInteractiveTerminal / parseCommaSeparatedList move from
  commands/utils.ts into lib/format.ts, re-exported for the ~50 command-layer
  callers -- the pattern utils.ts already used for VersionNotInstalledError.
  lib/drift-sync.ts, lib/refresh.ts and lib/secrets/list-filter.ts no longer
  import upward out of lib/.

- Docs: SESSION_AGENTS is 13, not 'exactly these 12' (muse was missing, in both
  specifications.md and architecture.md); sessions sync does not exist and never
  has, so it is removed from the command surface and from SES-IF-2; nine real
  subcommands that were missing are added.

Also commits the architecture scan this came from as a plan artifact.

Behavior-preserving except the byte-size rendering noted above.
