+++ apps/cli/docs/architecture.md
@@ -155,7 +155,7 @@ Two formats trip people up (both handled in [`src/lib/session/discover.ts`](../s
 the **session-discoverable** set is `SESSION_AGENTS`
 ([`src/lib/session/types.ts`](../src/lib/session/types.ts)): `claude`, `codex`,
 `gemini`, `antigravity`, `opencode`, `openclaw`, `rush`, `hermes`, `grok`, `kimi`,
-`droid`, `cursor`. `copilot` is runnable but not session-discoverable.
+`droid`, `cursor`, `muse`. `copilot` is runnable but not session-discoverable.
 `isSessionTrackedAgent()` in the same file is the single predicate every session-index
 writer gates on.
