+++ apps/cli/src/lib/secrets/list-filter.ts
@@ -18,7 +18,7 @@
  * and every axis narrows independently so they AND-compose.
  */
 
-import { parseCommaSeparatedList } from '../../commands/utils.js';
+import { parseCommaSeparatedList } from '../format.js';
 // From the leaf module, NOT discover.js — the latter imports `../sqlite.js`, so
 // reaching through it for one parser would pull node:sqlite into every `agents
 // secrets` invocation and print Node's SQLite ExperimentalWarning on stderr.
